% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch.R
\name{batch_results_preprocess}
\alias{batch_results_preprocess}
\title{Pre-process batch API data for use in this project}
\usage{
batch_results_preprocess(file_output_id)
}
\arguments{
\item{file_output_id}{Batch job output_file_id to process}
}
\value{
Data frame with batch info
}
\description{
Pre-process batch API data for use in this project
}
