% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch.R
\name{batch_score_process}
\alias{batch_score_process}
\title{Process batch scoring data}
\usage{
batch_score_process(batch_id, conn)
}
\arguments{
\item{batch_id}{Batch ID to process the competency scoring for}

\item{conn}{DB connection}
}
\value{
Data frame with batch info
}
\description{
Process batch scoring data
}
