% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{clear_cache}
\alias{clear_cache}
\title{Clear the cache of generated texts}
\usage{
clear_cache(log = FALSE)
}
\arguments{
\item{log}{If \code{TRUE}, also delete the call log.}
}
\value{
The number of cached texts removed, invisibly.
}
\description{
Clear the cache of generated texts
}
