% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbOperations.R
\name{dbAddCoreFaculty}
\alias{dbAddCoreFaculty}
\title{Add core faculty start dates and update core faculty status in evaluations}
\usage{
dbAddCoreFaculty(file, conn)
}
\arguments{
\item{file}{Path to a CSV file with columns original_evaluator_id and core_faculty_start}

\item{conn}{A database connection (schema inst/narrate.sql)}
}
\value{
TRUE invisibly on success
}
\description{
Add core faculty start dates and update core faculty status in evaluations
}
