% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbOperations.R
\name{dbAddPrompt}
\alias{dbAddPrompt}
\title{Add a new prompt to the database}
\usage{
dbAddPrompt(prompt, conn, note, task = NULL, commit = T, showWarning = T)
}
\arguments{
\item{prompt}{Single string of system prompt text}

\item{conn}{NARRATE database connection}

\item{note}{(Optional) Note about this prompt}

\item{commit}{(Default = TRUE) Commit the transaction}

\item{showWarning}{(Default = TRUE) Show warning if prompt already exists}
}
\value{
Prompt ID
}
\description{
Add a new prompt to the database
}
