% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbOperations.R
\name{dbCompExtraction}
\alias{dbCompExtraction}
\title{Insert or update a list of extracted competencies}
\usage{
dbCompExtraction(
  conn,
  review_assignment_id,
  comp_extraction,
  return_tables = F,
  commit = T
)
}
\arguments{
\item{conn}{CFME database connection}

\item{review_assignment_id}{Review assignment ID}

\item{comp_extraction}{List as generated by llm_comp_extract() (i.e. result$data),
where each element has a cID (integer) and text (character vector)}

\item{return_tables}{(Default = F) If TRUE, returns competency_score and
competency_text as dataframes in the result list}

\item{commit}{(Default = T)}
}
\value{
A list with success (T/F) and optionally competency_score and
competency_text dataframes if return_tables = TRUE
}
\description{
Insert or update a list of extracted competencies
}
