% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbOperations.R
\name{dbReviewUpdate}
\alias{dbReviewUpdate}
\title{Insert or update into review score table}
\usage{
dbReviewUpdate(
  conn,
  statusCode,
  overallScores,
  compScores,
  compText,
  removeNotListed = F,
  commit = T
)
}
\arguments{
\item{conn}{NARRATE database connection}

\item{statusCode}{Set the review_assignment status; run status_codes(conn, "review_assignment") for code details}

\item{overallScores}{Data frame matching review_assignment table which
contains the overall scores}

\item{compScores}{Data frame matching competency_scores table (new IDs will be generated)}

\item{compText}{Data frame matching competency_text table (new IDs will be generated)}

\item{commit}{(Default = TRUE) Commit the transaction}
}
\value{
A list with the updated results from the database
}
\description{
Insert or update into review score table
}
