% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/llm.R
\name{llm_batch_create}
\alias{llm_batch_create}
\title{Submit a batch job}
\usage{
llm_batch_create(
  file_input_id,
  endpoint = llm_default_endpoint,
  api_key = Sys.getenv("HUIT_API_NARRATE")
)
}
\arguments{
\item{file_input_id}{File ID from llm_batch_upload()}

\item{endpoint}{Default = llm_default_endpoint. Gateway base URL}

\item{api_key}{API key. Default = HUIT_API_NARRATE env var}
}
\value{
Batch ID string
}
\description{
Submit a batch job
}
