% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/review_helper.R
\name{llm_batch_status}
\alias{llm_batch_status}
\title{Check and update the status of a batch job}
\usage{
llm_batch_status(
  batch_id,
  conn,
  check = TRUE,
  endpoint = llm_default_endpoint,
  api_key = Sys.getenv("HUIT_API_NARRATE")
)
}
\arguments{
\item{batch_id}{Internal batch ID (row id in the batch table)}

\item{conn}{DB connection}

\item{check}{Query the API for an updated status. Default = TRUE}

\item{endpoint}{Gateway base URL. Default = llm_default_endpoint}

\item{api_key}{API key. Default = HUIT_API_NARRATE env var}
}
\value{
Updated batch info data frame
}
\description{
Queries the OpenAI batch API and updates the batch table in the database.
}
