% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_mod_highlight.R
\name{mod_highlight_ui_controls}
\alias{mod_highlight_ui_controls}
\title{Module UI for text highlight controls}
\usage{
mod_highlight_ui_controls(
  id,
  element,
  label = "Text evidence",
  info =
    "<i>Select pieces of text and click the 'Add highlighted' button to add them to the list below</i>"
)
}
\arguments{
\item{id}{Module ID}

\item{element}{(Optional) Limit text selection to a specific HTML element ID
and its children (don't provide the #)}

\item{label}{(Default = = "Text selections"). Label for the UI list}

\item{info}{(Has default) HTML to display underneath the label}
}
\value{
Shiny UI element
}
\description{
UI has a button with underneath a compact element where a list of selected
peices of text (belonging to the currently active group) are displayed and
can be removed again.
}
\details{
Must be paired with a \code{\link[=mod_highlight_ui_text]{mod_highlight_ui_text()}} call using the \emph{same}
module \code{id}, and both wired up to a single \code{\link[=mod_highlight_server]{mod_highlight_server()}} call
(also using that \code{id}).
}
