% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{pinDB}
\alias{pinDB}
\title{Backup and replace the DB using pins}
\usage{
pinDB(
  dbPath,
  action,
  exportPin = "narrate_db_export",
  importPin = "narrate_db_import",
  nBackups = 3
)
}
\arguments{
\item{dbPath}{Path to the DB}

\item{action}{Any of the following: "import", "export". Can be both as vector}

\item{exportPin}{(Default = "narrate_db_export") Pin name for the export / backup DB}

\item{importPin}{(Default = "narrate_db_import") Pin name for the import DB}

\item{nBackups}{(Default = 3) N most recent exports to keep}

\item{password}{Admin password, set \code{adminPass} as an environment variable}
}
\value{
list with success an msg
}
\description{
Backup and replace the DB using pins
}
