% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rubric.R
\name{rubric_parsing}
\alias{rubric_parsing}
\title{Parse inst/rubric.md and insert competency, disambiguation, and score data into the database}
\usage{
rubric_parsing(conn, rubric_path = NULL, commit = T)
}
\arguments{
\item{conn}{Database connection}

\item{rubric_path}{Path to rubric.md (defaults to package inst file)}

\item{commit}{(Default = TRUE) Commit the results to the database.}
}
\value{
Invisibly, a list of inserted data frames: competency, competency_diff, score
}
\description{
Parse inst/rubric.md and insert competency, disambiguation, and score data into the database
}
