% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rubric.R
\name{rubric_process}
\alias{rubric_process}
\title{Check the rubric file to any updates and reprocess to generate prompts if needed}
\usage{
rubric_process(conn, showWarning = F)
}
\arguments{
\item{conn}{Database connection}
}
\value{
Table with latest set of prompts
}
\description{
Check the rubric file to any updates and reprocess to generate prompts if needed
}
