% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{status_codes}
\alias{status_codes}
\title{Look up status codes for a database table or function}
\usage{
status_codes(conn, table = NULL)
}
\arguments{
\item{conn}{CFME database connection}

\item{table}{Name of a database table or function to filter by (optional)}
}
\value{
Data frame of matching status codes
}
\description{
Look up status codes for a database table or function
}
