% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch.R
\name{batch_extract_process}
\alias{batch_extract_process}
\title{Process batch extraction data}
\usage{
batch_extract_process(batch_id, conn)
}
\arguments{
\item{batch_id}{Batch ID to process the competency extraction for}

\item{conn}{DB connection}
}
\value{
Data frame with batch info
}
\description{
Process batch extraction data
}
