% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/review_helper.R
\name{db_write_score_specificity}
\alias{db_write_score_specificity}
\title{Update competency_score.specificity for a single review assignment}
\usage{
db_write_score_specificity(conn, rid, competencies, commit = FALSE)
}
\arguments{
\item{conn}{DB connection}

\item{rid}{review_assignment ID}

\item{competencies}{List of items with cID and specificity fields}

\item{commit}{Commit the transaction. Default = FALSE}
}
\description{
Used by both llm_comp_score_run() and batch_score_process().
}
