% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{deployShinyApp}
\alias{deployShinyApp}
\title{Delop Shiny App}
\usage{
deployShinyApp(db, gitHubBranch, dev = F, app_file = "inst/review_app.R")
}
\arguments{
\item{db}{Database to use}

\item{gitHubBranch}{NARRATE branch}

\item{dev}{Deploy to dev app}
}
\value{
Nothing
}
\description{
Delop Shiny App
}
