% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch.R
\name{llm_batch_create}
\alias{llm_batch_create}
\title{Submit a batch job}
\usage{
llm_batch_create(
  file_id,
  endpoint = "https://azure-ai.hms.edu",
  api_key = Sys.getenv("HMS_AZURE_API")
)
}
\arguments{
\item{file_id}{From llm_batch_upload()}
}
\value{
Batch ID string
}
\description{
Submit a batch job
}
