% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/llm.R
\name{llm_batch_create}
\alias{llm_batch_create}
\title{Submit a batch job}
\usage{
llm_batch_create(
  file_input_id,
  endpoint = "https://azure-ai.hms.edu",
  api_key = Sys.getenv("HMS_AZURE_API")
)
}
\arguments{
\item{file_input_id}{File ID from llm_batch_upload()}

\item{endpoint}{Azure endpoint base URL}

\item{api_key}{API key. Default = HMS_AZURE_API env var}
}
\value{
Batch ID string
}
\description{
Submit a batch job
}
