% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch.R
\name{llm_batch_status}
\alias{llm_batch_status}
\title{Poll a batch job until it reaches a terminal state}
\usage{
llm_batch_status(
  batch_id,
  conn,
  check = T,
  endpoint = "https://azure-ai.hms.edu",
  api_key = Sys.getenv("HMS_AZURE_API")
)
}
\arguments{
\item{batch_id}{Batch ID to check / process}

\item{conn}{DB connection}

\item{check}{Use the API to check the status if not finished yet}
}
\value{
Data frame with batch info
}
\description{
Poll a batch job until it reaches a terminal state
}
