% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/llm.R
\name{llm_batch_upload}
\alias{llm_batch_upload}
\title{Upload a JSONL file to the Azure OpenAI Files API}
\usage{
llm_batch_upload(
  jsonl_content,
  endpoint = "https://azure-ai.hms.edu",
  api_key = Sys.getenv("HMS_AZURE_API")
)
}
\arguments{
\item{jsonl_content}{Output of llm_batch_build_jsonl()}

\item{endpoint}{Azure endpoint base URL}

\item{api_key}{API key. Default = HMS_AZURE_API env var}
}
\value{
File ID string
}
\description{
Upload a JSONL file to the Azure OpenAI Files API
}
