% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/llm.R
\name{llm_chat_completion}
\alias{llm_chat_completion}
\title{Call the Azure OpenAI chat completions API}
\usage{
llm_chat_completion(
  user,
  system,
  log,
  model = "gpt-4o-1120",
  maxTokens = 500,
  version = "2024-10-21",
  endpoint = "https://azure-ai.hms.edu"
)
}
\arguments{
\item{user}{User prompt}

\item{system}{System prompt. Default = "You are a helpful AI assistant"}

\item{log}{If set, token usage is appended to this CSV file}

\item{model}{Default = gpt-4o-1120. Azure deployment name}

\item{maxTokens}{Default = 500. Maximum tokens to return}

\item{version}{Default = 2024-10-21. API version}

\item{endpoint}{Default = https://azure-ai.hms.edu. Azure endpoint base URL}
}
\value{
Parsed response list
}
\description{
Note that this function expects an environment variable HMS_AZURE_API
that contains the API key. You can set this up using
\code{Sys.setenv(HMS_AZURE_API = "API token here")}
}
